style: reload 정렬  style fix
